+++ dev/src/dev.clj
@@ -1,6 +1,7 @@
 (ns dev
   (:require [demo.config :refer [config]]
             [demo.system :as system]
+            [datastar.wow.deacon :as d*conn]
             [malli.dev :as malli.dev]))
 
 (defn start []
@@ -15,3 +16,9 @@
   (stop))
 
 (start)
+
+(comment
+  (def s (d*conn/store {:type :caffeine :scheduler true :duration-ms 15000}))
+  (d*conn/store! s ::test {:conn true})
+  (d*conn/connection s ::test)
+  )